+++ src/sanity/schemas/index.ts
@@ -1,5 +1,9 @@
 import { type SchemaTypeDefinition } from 'sanity'
+import { evenement   } from "./evenement"
+import { ressource   } from "./ressource"
+import { partenaire  } from "./partenaire"
+import { membreEquipe } from "./membreEquipe"
 
-export const schema: { types: SchemaTypeDefinition[] } = {
-  types: [],
+export const schema = {
+  types: [evenement, ressource, partenaire, membreEquipe],
 }